imageExperiments/requirements_exp.txt
absl-py==1.2.0
astunparse==1.6.3
bleach==5.0.1
cachetools==5.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
cryptography==38.0.1
cycler==0.11.0
docutils==0.19
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
jaraco.classes==3.2.2
jeepney==0.8.0
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.9.1
kiwisolver==1.4.4
libclang==14.0.6
lime==0.2.0.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
more-itertools==8.14.0
networkx==2.8.6
nltk==3.7
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pkginfo==1.8.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
readme-renderer==37.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==3.3.3
shap==0.41.0
six==1.16.0
slicer==0.0.7
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tqdm==4.64.1
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
